backend/requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
transformers==4.35.2
torch==2.3.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
gunicorn==21.2.0
Werkzeug==2.3.8
twilio==8.10.0
google-cloud-speech==2.31.1
